requirements.txt
django==1.11.12
django-crispy-forms==1.7.2
markdown==2.6.11
